GenerateNewKey et GenerateKeyFor.
- usage : "GenerateNewKey [amount]". Invente un nouveau KeyValue pour
l'item sélectionné et crée [amount] clefs, qui sont mise dans le
backpack du GM. On peut créer entre 0 et 10 clefs à la fois.

- usage : "GenerateKeyFor [amount]". Utilise le KeyValue de l'item
sélectionné, et crée [amount] clefs, qui sont mise dans le backpack du
GM. On peut créer entre 0 et 10 clefs à la fois.

+++ Scripts/Items/Misc/KeyRing.cs
@@ -143,7 +143,7 @@ namespace Server.Items
 			UpdateItemID();
 		}
 
-		public void RemoveKeys( uint keyValue )
+		public void RemoveKeys( long keyValue )
 		{
 			for ( int i = m_Keys.Count - 1; i >= 0; i-- )
 			{
@@ -159,7 +159,7 @@ namespace Server.Items
 			UpdateItemID();
 		}
 
-		public bool ContainsKey( uint keyValue )
+        public bool ContainsKey( long keyValue)
 		{
 			foreach ( Key key in m_Keys )
 			{
